Remove subsidy parameters (#2380)
# Description
Fixes https://github.com/cowprotocol/services/issues/2369

# Changes
Simplifies fees: removes all parameters for subsidies and all dependent
code.

## How to test
Existing tests.
